added newtycoon automation
Automation added to automatically add a new button into the UI, fully configured for use.
Please remember to fill out all variables to allow the script to function properly.
Expanding the frame is not yet done, feel free to create a branch and edit the code if you know how to fix it. I will work on it in the future since this is hindering my productivity.
For now, just add 120 to the value at game.StarterGui.ScreenGui.NavCatalog.Library.CanvasSize.Y.Offset